+++ app/canary/play.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useMemo, useRef, useCallback, memo } from 'react';
 import {
   View, Text, FlatList, TouchableOpacity, TextInput,
-  StyleSheet, useWindowDimensions, Linking, Modal, Platform,
+  useWindowDimensions, Linking, Modal, Platform,
   ActivityIndicator, Animated, ScrollView
 } from 'react-native';
 import { useRouter, useLocalSearchParams } from 'expo-router';
@@ -10,8 +10,12 @@ import Ionicons from '@expo/vector-icons/Ionicons';
 
 import { GlitchText } from '@/assets/components/GlitchText';
 import { Game } from '../../assets/components/Game';
-import { gamesData } from './media/games';
+import { gamesData } from '../../assets/constants/games';
 import { GameWall } from '@/assets/components/GameWall';
+
+import { styles, C, GENRE_FILTERS, VIBES } from '@/assets/constants/Theme';
+import { FilterPill } from '@/assets/components/FilterPill';
+import { HorizontalRow, ControlIcon, SectionHeader, GameWrapper, StatChip } from '@/assets/components/Wrappers';
 
 // --- TYPES ---
 interface GameType {
@@ -41,235 +45,7 @@ const VER_PATCHES = [
   "53qx9jfl", "g213gtyp", "exrkqldd", "5gc3ymeh", "oumntxws",
   "5rs46hkv", "otg548fh",
 ];
-const VER_INFO = { date: '30/4/26', text: '8.1.2', patch: VER_PATCHES[6] };
-
-const GENRE_FILTERS = [
-  { id: 'all', label: '✦ All', color: '#818cf8' },
-  { id: 'action', label: '⚡ Action', color: '#f59e0b' },
-  { id: 'racing', label: '🏎 Racing', color: '#f87171' },
-  { id: 'sports', label: '🏆 Sports', color: '#fbbf24' },
-  { id: 'puzzle', label: '🧩 Puzzle', color: '#34d399' },
-  { id: 'survival', label: '💀 Survival', color: '#a78bfa' },
-  { id: 'adventure', label: '🗺 Adventure', color: '#fb923c' },
-  { id: 'clicker', label: '🖱 Clicker', color: '#e879f9' },
-  { id: 'platformer', label: '🏷️ Platformer', color: '#ef4444' },
-];
-
-// 46 different "vibe" texts.
-const VIBES = [
-  "no teachers here 😈",
-  "zero homework zone 🎮",
-  "skill issue? skill issue.",
-  "rage quit? never heard of it.",
-  "built different, plays different",
-  "lunch break? more like boss fight.",
-  "loading infinite fun...",
-  "attendance? never heard of her.",
-  "your high score is someone else's tutorial.",
-  "W only. no L's allowed.",
-  "this is not a drill. this is a boss.",
-  "we don't pause here.",
-  "npc behaviour detected. not you though.",
-  "undefeated since forever 🏆",
-  "touch grass later.",
-  "main character mode: activated.",
-  "the game doesn't end. you do.",
-  "lwk goated fr.",
-  "no cap, this hits different.",
-  "notifications off. game on.",
-  "respawn incoming... just kidding, you're fine.",
-  "teacher left the room 👀",
-  "speedrun any%",
-  "plot twist: you win.",
-  "ctrl + z doesn't work in real life. here it does.",
-  "silent but deadly at this game.",
-  "not procrastinating. practising.",
-  "the tutorial was optional. you're not.",
-  "your focus just unlocked. use it.",
-  "lag is a myth. you're just early.",
-  "zero gravity on the leaderboard rn.",
-  "certified banger alert 🔊",
-  "play stupid games, win stupid prizes... jk you always win.",
-  "wifi strong. excuses weak.",
-  "one more level. always one more level.",
-  "the grind doesn't stop, it just loads.",
-  "nobody beats you here. facts.",
-  "no homework was harmed in the making of this session.",
-  "free period loading... complete ✅",
-  "finals? we don't do that here.",
-  "high score or high standards? both.",
-  "living rent free in the leaderboard.",
-  "they said it couldn't be done. they were wrong.",
-  "game face: on. everything else: off.",
-  "extra life acquired 🟢",
-  "sleep is a checkpoint, not a game over.",
-  "insert coin. wait, it's free. go wild.",
-];
-
-// --- PALETTE ---
-const C = {
-  bg: '#020617',
-  surface: '#0a0f1e',
-  card: '#0d1526',
-  border: '#1a2744',
-  accent: '#6366f1',
-  accentLt: '#818cf8',
-  hot: '#f43f5e',
-  gold: '#fbbf24',
-  muted: '#475569',
-  mutedLt: '#64748b',
-  text: '#f1f5f9',
-  textDim: '#94a3b8',
-  green: '#34d399',
-};
-
-// --- ANIMATED PILL ---
-const FilterPill = memo(({ item, active, onPress }: any) => {
-  const scale = useRef(new Animated.Value(1)).current;
-
-  const handlePress = () => {
-    Animated.sequence([
-      Animated.timing(scale, { toValue: 0.92, duration: 80, useNativeDriver: true }),
-      Animated.timing(scale, { toValue: 1, duration: 120, useNativeDriver: true }),
-    ]).start();
-    onPress(item.id);
-  };
-
-  return (
-    <Animated.View style={{ transform: [{ scale }] }}>
-      <TouchableOpacity
-        onPress={handlePress}
-        activeOpacity={0.85}
-        style={[
-          styles.pill,
-          active && { backgroundColor: item.color + '22', borderColor: item.color },
-        ]}
-      >
-        <Text style={[styles.pillText, active && { color: item.color }]}>{item.label}</Text>
-      </TouchableOpacity>
-    </Animated.View>
-  );
-});
-
-// --- TRENDING BADGE ---
-const TrendingBadge = memo(() => (
-  <View style={styles.trendingBadge}>
-    <Text style={styles.trendingText}>🔥</Text>
-  </View>
-));
-
-// --- FAV BUTTON ---
-const FavButton = memo(({ isFav, onToggle }: any) => {
-  const scale = useRef(new Animated.Value(1)).current;
-
-  const handlePress = () => {
-    Animated.sequence([
-      Animated.timing(scale, { toValue: 1.4, duration: 100, useNativeDriver: true }),
-      Animated.timing(scale, { toValue: 1, duration: 150, useNativeDriver: true }),
-    ]).start();
-    onToggle();
-  };
-
-  return (
-    <TouchableOpacity onPress={handlePress} style={styles.favBadge} activeOpacity={0.8}>
-      <Animated.View style={{ transform: [{ scale }] }}>
-        <Ionicons name={isFav ? 'heart' : 'heart-outline'} size={15} color={isFav ? C.hot : '#fff'} />
-      </Animated.View>
-    </TouchableOpacity>
-  );
-});
-
-// --- GAME CARD ---
-const GameWrapper = memo(({ game, width, onPress, isFav, onToggleFav }: any) => {
-  return (
-    <View style={{ width, padding: 5 }}>
-      <View style={styles.gameCard}>
-        <Game
-          name={game.title.en}
-          imageSource={game.img}
-          broken={game.broken}
-          issueId={game.issue}
-          onPress={() => onPress(game)}
-        />
-        {game.popular && <TrendingBadge />}
-        <FavButton isFav={isFav} onToggle={() => onToggleFav(game.title.en)} />
-      </View>
-    </View>
-  );
-});
-
-// --- ICON BUTTON ---
-const ControlIcon = memo(({ name, onPress, color = C.accentLt, size = 21, active = false, label }: any) => (
-  <TouchableOpacity
-    onPress={onPress}
-    activeOpacity={0.7}
-    style={[styles.iconBtn, active && { backgroundColor: color + '25', borderColor: color + '80' }]}
-  >
-    <Ionicons name={name} size={size} color={active ? color : C.muted} />
-    {label ? <Text style={[styles.iconLabel, active && { color }]}>{label}</Text> : null}
-  </TouchableOpacity>
-));
-
-// --- STAT CHIP ---
-const StatChip = memo(({ value, label }: any) => (
-  <View style={styles.statChip}>
-    <Text style={styles.statValue}>{value}</Text>
-    <Text style={styles.statLabel}>{label}</Text>
-  </View>
-));
-
-// --- SECTION HEADER (optionally collapsible) ---
-const SectionHeader = memo(({ title, count, collapsed, onToggle }: any) => {
-  const rotate = useRef(new Animated.Value(collapsed ? 1 : 0)).current;
-
-  useEffect(() => {
-    Animated.timing(rotate, {
-      toValue: collapsed ? 1 : 0,
-      duration: 200,
-      useNativeDriver: true,
-    }).start();
-  }, [collapsed]);
-
-  const chevronRotate = rotate.interpolate({ inputRange: [0, 1], outputRange: ['0deg', '-90deg'] });
-
-  return (
-    <TouchableOpacity
-      onPress={onToggle}
-      activeOpacity={onToggle ? 0.7 : 1}
-      style={styles.sectionRow}
-      disabled={!onToggle}
-    >
-      <Text style={styles.sectionTitle}>{title}</Text>
-      {count !== undefined && (
-        <View style={styles.countBadge}>
-          <Text style={styles.countText}>{count}</Text>
-        </View>
-      )}
-      {onToggle && (
-        <Animated.View style={{ marginLeft: 'auto', transform: [{ rotate: chevronRotate }] }}>
-          <Ionicons name="chevron-down" size={18} color={C.muted} />
-        </Animated.View>
-      )}
-    </TouchableOpacity>
-  );
-});
-
-// --- HORIZONTAL GAME ROW ---
-const HorizontalRow = memo(({ games, onPress, favorites, onToggleFav }: any) => (
-  <ScrollView horizontal showsHorizontalScrollIndicator={false} style={{ marginHorizontal: -16 }} contentContainerStyle={{ paddingHorizontal: 16, gap: 10 }}>
-    {games.map((game: GameType) => (
-      <View key={game.title.en} style={styles.hCard}>
-        <Game
-          name={game.title.en}
-          imageSource={game.img}
-          broken={game.broken}
-          onPress={() => onPress(game)}
-        />
-        <FavButton isFav={favorites.includes(game.title.en)} onToggle={() => onToggleFav(game.title.en)} />
-      </View>
-    ))}
-  </ScrollView>
-));
+const VER_INFO = { date: '3/5/26', text: '8.1.3', patch: VER_PATCHES[9] };
 
 // --- MAIN SCREEN ---
 export default function HomeScreen() {
@@ -334,10 +110,10 @@ export default function HomeScreen() {
   // COMPUTED
   const columns = useMemo(() => {
     if (width < 480) return 2;
-    if (width < 768) return 3;
-    if (width < 1100) return 4;
-    if (width < 1400) return 6;
-    return 8;
+    if (width < 768) return 4;
+    if (width < 1100) return 6;
+    if (width < 1400) return 8;
+    return 10;
   }, [width]);
 
   const itemWidth = useMemo(() => (width - 32) / columns, [width, columns]);
@@ -585,325 +361,3 @@ export default function HomeScreen() {
     </View>
   );
 }
-
-// --- STYLES ---
-const styles = StyleSheet.create({
-  container: { flex: 1, backgroundColor: C.bg },
-  listPadding: { padding: 16, paddingBottom: 120 },
-  headerContainer: { marginBottom: 8 },
-
-  // HERO
-  hero: {
-    padding: 28,
-    borderRadius: 28,
-    backgroundColor: C.surface,
-    borderWidth: 1,
-    borderColor: C.border,
-    marginBottom: 20,
-    alignItems: 'center',
-  },
-  heroBadge: {
-    backgroundColor: C.hot + '20',
-    borderRadius: 100,
-    paddingHorizontal: 12,
-    paddingVertical: 4,
-    borderWidth: 1,
-    borderColor: C.hot + '60',
-    marginBottom: 14,
-  },
-  heroBadgeText: {
-    color: C.hot,
-    fontSize: 11,
-    fontWeight: '800',
-    letterSpacing: 2,
-  },
-  heroTitle: {
-    fontSize: 42,
-    fontWeight: '900',
-    color: C.accentLt,
-    letterSpacing: -1,
-    textAlign: 'center',
-  },
-  heroVibe: {
-    color: C.mutedLt,
-    fontSize: 13,
-    fontWeight: '500',
-    textAlign: 'center',
-    marginTop: 6,
-    fontStyle: 'italic',
-  },
-  statsRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginTop: 20,
-    marginBottom: 20,
-    gap: 8,
-  },
-  statChip: {
-    alignItems: 'center',
-    paddingHorizontal: 16,
-  },
-  statValue: {
-    color: C.text,
-    fontSize: 22,
-    fontWeight: '800',
-    letterSpacing: -0.5,
-  },
-  statLabel: {
-    color: C.muted,
-    fontSize: 11,
-    fontWeight: '600',
-    letterSpacing: 0.5,
-    textTransform: 'uppercase',
-    marginTop: 2,
-  },
-  statDivider: {
-    width: 1,
-    height: 32,
-    backgroundColor: C.border,
-  },
-  navRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    flexWrap: 'wrap',
-    justifyContent: 'center',
-  },
-  iconBtn: {
-    paddingHorizontal: 12,
-    paddingVertical: 8,
-    borderRadius: 14,
-    backgroundColor: C.card,
-    borderWidth: 1,
-    borderColor: C.border,
-    alignItems: 'center',
-    gap: 4,
-    minWidth: 52,
-  },
-  iconLabel: {
-    color: C.muted,
-    fontSize: 10,
-    fontWeight: '600',
-    textTransform: 'uppercase',
-    letterSpacing: 0.3,
-  },
-  vPipe: { width: 1, height: 32, backgroundColor: C.border },
-  verText: {
-    color: C.muted,
-    fontSize: 11,
-    fontWeight: '500',
-    marginTop: 16,
-    letterSpacing: 0.5,
-  },
-
-  // SEARCH
-  searchRow: { flexDirection: 'row', gap: 10, marginBottom: 16 },
-  searchRowFocused: {},
-  searchBox: {
-    flex: 1,
-    height: 52,
-    backgroundColor: C.surface,
-    borderRadius: 18,
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderColor: C.border,
-  },
-  searchBoxFocused: {
-    borderColor: C.accent + '80',
-    backgroundColor: C.card,
-  },
-  searchInput: {
-    flex: 1,
-    color: C.text,
-    fontSize: 15,
-    paddingHorizontal: 12,
-    fontWeight: '500',
-  },
-  randomBtn: {
-    width: 52,
-    height: 52,
-    backgroundColor: C.accent,
-    borderRadius: 18,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-
-  // GENRE PILLS
-  pillsScroll: { marginBottom: 20, marginHorizontal: -16 },
-  pillsContent: { paddingHorizontal: 16, gap: 8, flexDirection: 'row' },
-  pill: {
-    paddingHorizontal: 14,
-    paddingVertical: 8,
-    borderRadius: 100,
-    backgroundColor: C.card,
-    borderWidth: 1,
-    borderColor: C.border,
-  },
-  pillText: {
-    color: C.textDim,
-    fontSize: 13,
-    fontWeight: '600',
-  },
-
-  // SECTIONS
-  section: { marginBottom: 24 },
-  sectionRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 10,
-    marginBottom: 14,
-  },
-  sectionTitle: {
-    color: C.text,
-    fontSize: 18,
-    fontWeight: '800',
-    letterSpacing: -0.3,
-  },
-  countBadge: {
-    backgroundColor: C.accent + '25',
-    borderRadius: 100,
-    paddingHorizontal: 10,
-    paddingVertical: 3,
-    borderWidth: 1,
-    borderColor: C.accent + '50',
-  },
-  countText: {
-    color: C.accentLt,
-    fontSize: 12,
-    fontWeight: '700',
-  },
-
-  // GAME CARD
-  gameCard: {
-    borderRadius: 16,
-    overflow: 'hidden',
-    backgroundColor: C.card,
-    borderWidth: 1,
-    borderColor: C.border,
-    position: 'relative',
-  },
-  trendingBadge: {
-    position: 'absolute',
-    top: 8,
-    left: 8,
-    backgroundColor: 'rgba(0,0,0,0.75)',
-    borderRadius: 8,
-    paddingHorizontal: 6,
-    paddingVertical: 3,
-    borderWidth: 1,
-    borderColor: C.hot + '60',
-  },
-  trendingText: { fontSize: 12 },
-  favBadge: {
-    position: 'absolute',
-    top: 8,
-    right: 8,
-    backgroundColor: 'rgba(0,0,0,0.7)',
-    padding: 6,
-    borderRadius: 10,
-    borderWidth: 1,
-    borderColor: 'rgba(255,255,255,0.12)',
-  },
-
-  // HORIZONTAL CARD
-  hCard: {
-    width: 140,
-    borderRadius: 14,
-    overflow: 'hidden',
-    backgroundColor: C.card,
-    borderWidth: 1,
-    borderColor: C.border,
-    position: 'relative',
-  },
-
-  // EMPTY STATE
-  emptyState: { alignItems: 'center', paddingVertical: 60 },
-  emptyIcon: { fontSize: 48, marginBottom: 12 },
-  emptyTitle: { color: C.text, fontSize: 20, fontWeight: '800', marginBottom: 6 },
-  emptySubtitle: { color: C.muted, fontSize: 14, fontWeight: '500' },
-
-  // MODAL
-  modalBg: { flex: 1, backgroundColor: 'rgba(2,6,23,0.97)' },
-  modalContent: {
-    flex: 1,
-    margin: Platform.OS === 'web' ? 16 : 0,
-    borderRadius: 24,
-    overflow: 'hidden',
-    backgroundColor: C.surface,
-    borderWidth: 1,
-    borderColor: C.border,
-  },
-  modalBar: {
-    height: 64,
-    backgroundColor: C.card,
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingHorizontal: 16,
-    borderBottomWidth: 1,
-    borderBottomColor: C.border,
-    gap: 12,
-  },
-  closeBtn: {
-    width: 40,
-    height: 40,
-    borderRadius: 12,
-    backgroundColor: C.surface,
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderColor: C.border,
-  },
-  modalMeta: { flex: 1, gap: 3 },
-  modalTitle: {
-    color: C.text,
-    fontWeight: '800',
-    fontSize: 16,
-    letterSpacing: -0.3,
-  },
-  modalTrendChip: {
-    alignSelf: 'flex-start',
-    backgroundColor: C.hot + '20',
-    borderRadius: 6,
-    paddingHorizontal: 8,
-    paddingVertical: 2,
-    borderWidth: 1,
-    borderColor: C.hot + '50',
-  },
-  modalTrendText: {
-    color: C.hot,
-    fontSize: 10,
-    fontWeight: '700',
-    letterSpacing: 0.3,
-  },
-  modalActions: { flexDirection: 'row', gap: 8 },
-  modalIconBtn: {
-    width: 38,
-    height: 38,
-    borderRadius: 12,
-    backgroundColor: C.surface,
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderColor: C.border,
-  },
-  modalIconActive: {
-    borderColor: C.hot + '60',
-    backgroundColor: C.hot + '15',
-  },
-  loaderContainer: {
-    position: 'absolute',
-    top: 0, left: 0, right: 0, bottom: 0,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: C.bg,
-    gap: 12,
-    zIndex: 10,
-  },
-  loaderText: {
-    color: C.muted,
-    fontSize: 13,
-    fontWeight: '600',
-    fontStyle: 'italic',
-  },
-});
