--- app/devpatch/media/games.ts
@@ -1,90 +1,0 @@
-export const gamesData = [
-  { title: { en: "OvO" }, img: "7", url: "/ovo/", genre: "platformer" },
-  { title: { en: "OvO 2" }, img: "bh", url: "/ovo2.html", genre: "platformer", pc: true },
-  { title: { en: "OvO Dimensions" }, img: "be", url: "/ovo-dimensions.html", genre: "platformer" },
-  { title: { en: "Stickman Hook" }, img: "ax", url: "/stickman-hook.html", genre: "platformer" },
-  { title: { en: "Ragdoll Archers" }, img: "2", url: "/ragdoll-archers/", genre: "action" },
-  { title: { en: "Ragdoll Hit" }, img: "c", url: "/ragdoll-hit/", genre: "action" },
-  { title: { en: "Recoil" }, img: "bf", url: "/recoil.html", genre: "action" },
-  { title: { en: "GunSpin" }, img: "8", url: "/gunspin/", genre: "action" },
-  { title: { en: "Knife Hit" }, img: "bg", url: "/knifehit/", genre: "puzzle" },
-  { title: { en: "Slice Master" }, img: "q", url: "https://game-hub.nyc3.cdn.digitaloceanspaces.com/slice-master/index.html", genre: "puzzle", broken: true, issue: '60' },
-  { title: { en: "Fruit Ninja" }, img: "v", url: "/fruitninja.htm", genre: "puzzle" },
-  { title: { en: "Fast Runner" }, img: "t", url: "/fast-runner/", genre: "racing" },
-  { title: { en: "Swoop" }, img: "ap", url: "/SWOOOP.html", genre: "action" },
-  { title: { en: "Jetpack Joyride" }, img: "aw", url: "/jetpack-joyride/", genre: "action" },
-  { title: { en: "Tomb of the Mask" }, img: "ae", url: "/totm.html", genre: "action" },
-  { title: { en: "Subway Surfers" }, img: "3", url: "/subway-surfers/", genre: "platformer" },
-  { title: { en: "Roper" }, img: "b", url: "/roper/", genre: "action" },
-  { title: { en: "Thorns and Balloons" }, img: "5", url: "/thorns-and-balloons/", genre: "puzzle" },
-  { title: { en: "Moto X3M" }, img: "bl", url: "/moto-x3m.html", genre: "racing" },
-  { title: { en: "Moto X3M Winter" }, img: "0", url: "/x3m/", genre: "racing" },
-  { title: { en: "Drive Mad" }, img: "9", url: "/drive-mad/", genre: "racing" },
-  { title: { en: "Drift Boss" }, img: "u", url: "/drift-boss/", genre: "racing" },
-  { title: { en: "Driven Wild" }, img: "ao", url: "/drivenwild/index.html", genre: "racing" },
-  { title: { en: "Monster Tracks" }, img: "o", url: "/monster-tracks.html", genre: "racing" },
-  { title: { en: "Car Drawing Game" }, img: "ad", url: "/car-drawing-game/", genre: "racing" },
-  { title: { en: "Formula Rush" }, img: "bw", url: "/formula-rush/", genre: "racing" },
-  { title: { en: "Battle Racing Stars" }, img: "ca", url: "/battle-racing-stars/", genre: "racing", broken: true },
-  { title: { en: "Sky Race 3D" }, img: "bk", url: "/skyrace3d.html", genre: "racing" },
-  { title: { en: "Wheelie Bike" }, img: "l", url: "/wheelie-bike/", genre: "racing" },
-  { title: { en: "Crossy Road" }, img: "ba", url: "/crossy-road.html", genre: "racing" },
-  { title: { en: "Cut the Rope" }, img: "ay", url: "/ctr/", genre: "puzzle" },
-  { title: { en: "Cut The Rope: Time Travel" }, img: "bt", url: "/time-travel/index.html", genre: "puzzle" },
-  { title: { en: "Happy Glass" }, img: "an", url: "/happyglass/", genre: "puzzle" },
-  { title: { en: "Nut Sort" }, img: "aj", url: "/nutsort/", genre: "puzzle" },
-  { title: { en: "Liquid Sort" }, img: "by", url: "/liquid-sort/", genre: "puzzle" },
-  { title: { en: "99 Balls" }, img: "cb", url: "/ninety-nine-balls/", genre: "puzzle" },
-  { title: { en: "Block Blast" }, img: "aa", url: "/block-blast/", genre: "puzzle" },
-  { title: { en: "Stack" }, img: "bz", url: "/stack/", genre: "puzzle" },
-  { title: { en: "Roll" }, img: "w", url: "/roll.html", genre: "puzzle" },
-  { title: { en: "Draw Climber" }, img: "g", url: "/draw-climber/", genre: "puzzle" },
-  { title: { en: "Draw the Hill" }, img: "bc", url: "/draw-the-hill/", genre: "puzzle" },
-  { title: { en: "Spiral Roll" }, img: "i", url: "/slice-roll/", genre: "action" },
-  { title: { en: "There Is No Game" }, img: "af", url: "/there-is-no-game.html", genre: "puzzle" },
-  { title: { en: "Impossible Quiz" }, img: "bj", url: "/impossiblequiz/", genre: "puzzle" },
-  { title: { en: "Hundred Percent Golf" }, img: "bv", url: "/hundred-percent-golf/index.html", genre: "sports" },
-  { title: { en: "Small World Cup" }, img: "ag", url: "/small-world-cup.html", genre: "sports" },
-  { title: { en: "Basket Random" }, img: "ac", url: "/randoms/basket/", genre: "sports" },
-  { title: { en: "Boxing Random" }, img: "ab", url: "/randoms/boxing/", genre: "sports" },
-  { title: { en: "Soccer Random" }, img: "ak", url: "/randoms/soccer/", genre: "sports" },
-  { title: { en: "Volley Random" }, img: "al", url: "/randoms/volley/", genre: "sports" },
-  { title: { en: "Penalty Kick Online" }, img: "e", url: "/penkick/", genre: "sports" },
-  { title: { en: "Tap Goal" }, img: "s", url: "/tap-goal/", genre: "sports" },
-  { title: { en: "Idle Football Manager" }, img: "k", url: "/idle-football/", genre: "sports" },
-  { title: { en: "Rocket Goal" }, img: "bq", url: "/rocketgoal/", genre: "sports" },
-  { title: { en: "Darts Pro" }, img: "f", url: "/dartspro/", genre: "sports" },
-  { title: { en: "Google Games Baseball" }, img: "y", url: "/baseball/", genre: "sports" },
-  { title: { en: "Google Games Cricket" }, img: "z", url: "/cricket17/", genre: "sports" },
-  { title: { en: "Bloons Tower Defence 5" }, img: "m", url: "/btd5.htm", genre: "action" },
-  { title: { en: "Plants vs Zombies" }, img: "p", url: "/pvz/", genre: "action" },
-  { title: { en: "Angry Birds" }, img: "bs", url: "/angrybirds.htm", genre: "action" },
-  { title: { en: "Survival Race" }, img: "d", url: "/survival-race/", genre: "sports" },
-  { title: { en: "No Pain, No Gain" }, img: "bu", url: "/npng/", genre: "action" },
-  { title: { en: "Bridge Race" }, img: "bp", url: "/bridge-race/", genre: "survival" },
-  { title: { en: "Rise Higher" }, img: "bx", url: "/rise-higher/", genre: "survival" },
-  { title: { en: "Flappy Bird" }, img: "h", url: "/flappy-bird/", genre: "survival" },
-  { title: { en: "Crazy Crash Landing" }, img: "n", url: "/ccl.htm", genre: "survival" },
-  { title: { en: "Geometry Dash Wave" }, img: "x", url: "/gd-wave/", genre: "survival" },
-  { title: { en: "Geometry Dash SubZero" }, img: "j", url: "/gd-sub0.htm", genre: "survival" },
-  { title: { en: "Geometry Dash World" }, img: "av", url: "/gdworld.htm", genre: "survival" },
-  { title: { en: "Duck Duck Clicker" }, img: "4", url: "/duck-duck-clicker/", genre: "clicker" },
-  { title: { en: "Cookie Clicker" }, img: "bo", url: "/cookieclicker/", genre: "clicker", broken: true, issue: '59' },
-  { title: { en: "Doge Miner" }, img: "bi", url: "/DogeMiner/", genre: "clicker" },
-  { title: { en: "Doge Miner 2" }, img: "bd", url: "/doge-miner-2/", genre: "clicker" },
-  { title: { en: "Monkey Mart" }, img: "bb", url: "/monkey-mart.html", genre: "adventure" },
-  { title: { en: "BitLife" }, img: "6", url: "/bitlife/", genre: "adventure" },
-  { title: { en: "Gobble" }, img: "r", url: "/gobble/", genre: "action" },
-  { title: { en: "Paper.io 2" }, img: "am", url: "/paperio2/", genre: "survival" },
-  { title: { en: "Tiny Fishing" }, img: "1", url: "/tiny-fishing/", genre: "clicker" },
-  { title: { en: "NEW Tiny Fishing" }, img: "aq", url: "/tiny-fishing/update", genre: "clicker" },
-  { title: { en: "Breaking the Bank" }, img: "az", url: "/stickmin/breakingthebank/", genre: "puzzle" },
-  { title: { en: "Stealing the Diamond" }, img: "at", url: "/stickmin/stealingthediamond/", genre: "puzzle" },
-  { title: { en: "Escaping the Prison" }, img: "ar", url: "/stickmin/escapingtheprison/", genre: "puzzle" },
-  { title: { en: "Fleeing the Complex" }, img: "as", url: "/stickmin/fleeingthecomplex/", genre: "puzzle" },
-  { title: { en: "Infiltrating the Airship" }, img: "au", url: "/stickmin/infiltratingtheairship/", genre: "puzzle" },
-  { title: { en: "Five Nights at Freddy's 1" }, img: "a", url: "/fnaf/FNAF1/", genre: "survival", horror: true },
-  { title: { en: "Five Nights at Freddy's 2" }, img: "ah", url: "/fnaf/FNAF2/", genre: "survival", horror: true },
-  { title: { en: "Five Nights at Freddy's 3" }, img: "ai", url: "/fnaf/FNAF3/", genre: "survival", horror: true },
-  { title: { en: "Five Nights at Epstein's" }, img: "br", url: "/fnae.html", genre: "survival", horror: true },
-];
